+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 0.13.0
 
+* Adding the `clustering/key-collision` namespace.
 * Adding the `clustering/naive` namespace.
+* Adding the `clustering/vp-tree` namespace.
 * Adding the `metrics/distance/identity` namespace.
 * Adding the `metrics/distance/monge-elkan` namespace.